requirements.txt
openai
replicate
requests
python-dotenv
pillow 